Fix English errors and make language clearer
This file was already well-written :) . I simplified language in order to make it easier to understand and translate.

There is still one sentence that I don't understand at lines 115-116.

Also, the term "mini computer" makes sense but is not as precise as "small single-board computer" (see the Wikipedia English pages for Raspberry Pi and for minicomputer). Maybe more accurate in this case would be "low-power computer"?